pyproject.toml
[project]
name = "reprod-interview-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "av>=16.0.1",
    "black>=25.12.0",
    "datasets>=4.4.2",
    "fasttext>=0.9.3",
    "hydra-core>=1.3.2",
    "ipython>=9.8.0",
    "lightning>=2.6.0",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchcodec>=0.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
]
